requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pydantic>=2.9.2
pydantic-settings>=2.5.2
ebooklib>=0.18
beautifulsoup4>=4.12.3
lxml>=5.3.0
anthropic>=0.39.0
httpx>=0.27.2
zeroconf>=0.135.0
pydub>=0.25.1
python-dotenv>=1.0.1
